+++ NEWS.md
@@ -1,9 +1,26 @@
+# nichetools 0.3.0
+
+### 22-August-2024
+
+*`extract_similiarties()`, `create_comparisons()`, `extract_group_metric()` 
+as well as additional arguments have been added to `extract_layman()` to extract
+either maximum-likelihood or Bayesian estimates. 
+
+* Test for all new functions have been added.
+
+* {SIBER} vignette is now completed
+
+# nichetools 0.2.1
+
+### 21-June-2024
+* `extract_layman` has been implemented 
+*  Tests have been added for this function 
+
 # nichetools 0.2.0
 
 ### 03-June-2024
 * functions now fully support {SIBER}
 * functions now extract randomly sampled ellipse 
-
 
 ### 16-May-2024
 * Update to documentation for CRAN submission